+++ src/nowiki/doubleUnderscore.ts
@@ -21,6 +21,7 @@ import {syntax} from '../../mixin/syntax';
 @hiddenToken() @padded('__')
 export abstract class DoubleUnderscoreToken extends NowikiBaseToken {
 	declare readonly name: string;
+	readonly #fullWidth;
 
 	/* NOT FOR BROWSER */
 
@@ -35,12 +36,14 @@ export abstract class DoubleUnderscoreToken extends NowikiBaseToken {
 	/**
 	 * @param word 状态开关名
 	 * @param sensitive 是否固定大小写
+	 * @param fullWidth 是否为全角下划线
 	 */
-	constructor(word: string, sensitive: boolean, config: Config, accum?: Token[]) {
+	constructor(word: string, sensitive: boolean, fullWidth: boolean, config: Config, accum?: Token[]) {
 		super(word, config, accum);
 		const lc = word.toLowerCase(),
 			{doubleUnderscore: [,, iAlias, sAlias]} = config;
 		this.setAttribute('name', (sensitive ? sAlias?.[word]?.toLowerCase() : iAlias?.[lc]) ?? lc);
+		this.#fullWidth = fullWidth;
 
 		/* NOT FOR BROWSER */
 
@@ -50,20 +53,24 @@ export abstract class DoubleUnderscoreToken extends NowikiBaseToken {
 
 	/** @private */
 	override toString(): string {
-		return `__${this.innerText}__`;
+		const underscore = this.#fullWidth ? '＿＿' : '__';
+		return underscore + this.innerText + underscore;
 	}
 
 	/** @private */
 	override print(): string {
-		return super.print({pre: '__', post: '__'});
+		const underscore = this.#fullWidth ? '＿＿' : '__';
+		return super.print({pre: underscore, post: underscore});
 	}
 
 	/* NOT FOR BROWSER */
 
 	override cloneNode(): this {
 		const config = this.getAttribute('config');
-		// @ts-expect-error abstract class
-		return Shadow.run((): this => new DoubleUnderscoreToken(this.innerText, this.#sensitive, config));
+		return Shadow.run(
+			// @ts-expect-error abstract class
+			(): this => new DoubleUnderscoreToken(this.innerText, this.#sensitive, this.#fullWidth, config),
+		);
 	}
 }